move xml-stripping functions into helpers
